requirements.txt
openai
python-dotenv
pandas
pyyaml
